bakery/requirements.txt
Django==3.9.0
django-debug-toolbar==3.1.1
djangorestframework==3.13.1
requests==2.24.0